refactor: Split autosar_models.py into domain-based modules
Refactored the monolithic autosar_models.py file (814 lines) into multiple
domain-based Python files to improve code organization and maintainability.

New file structure:
- enums.py: ATPType, AttributeKind
- base.py: AbstractAutosarBase
- attributes.py: AutosarAttribute, AutosarEnumLiteral
- types.py: AutosarClass, AutosarEnumeration, AutosarPrimitive
- containers.py: AutosarPackage, AutosarDoc

Key improvements:
- Domain-based organization by function
- Clear dependency hierarchy with no circular imports
- Improved maintainability and code navigation
- Full backward compatibility maintained

Quality checks:
- ✅ All 135 model unit tests pass (100%)
- ✅ Mypy type checking passes with no issues
- ✅ Ruff linting passes with no errors

Closes #70
