+++ scripts/report/coverage.md
@@ -2,13 +2,13 @@
 COVERAGE REPORT - MARKDOWN FORMAT
 ======================================================================
 
-## Overall Coverage: 86.9%
+## Overall Coverage: 92.4%
 
-**Total Statements**: 849
+**Total Statements**: 878
 
-**Statements Covered**: 738
+**Statements Covered**: 811
 
-**Statements Missing**: 111
+**Statements Missing**: 67
 
 ## All Source Files Coverage
 
@@ -17,10 +17,14 @@ COVERAGE REPORT - MARKDOWN FORMAT
 | ✓ `src\autosar_pdf2txt\__init__.py` | 5 | 5 | 0 | 100.0% |
 | ✓ `src\autosar_pdf2txt\cli\__init__.py` | 2 | 2 | 0 | 100.0% |
 | ✗ `src\autosar_pdf2txt\cli\autosar_cli.py` | 83 | 67 | 16 | 80.7% |
-| ✓ `src\autosar_pdf2txt\models\__init__.py` | 2 | 2 | 0 | 100.0% |
-| ✗ `src\autosar_pdf2txt\models\autosar_models.py` | 180 | 168 | 12 | 93.3% |
+| ✓ `src\autosar_pdf2txt\models\__init__.py` | 6 | 6 | 0 | 100.0% |
+| ✓ `src\autosar_pdf2txt\models\attributes.py` | 34 | 34 | 0 | 100.0% |
+| ✓ `src\autosar_pdf2txt\models\base.py` | 14 | 14 | 0 | 100.0% |
+| ✗ `src\autosar_pdf2txt\models\containers.py` | 96 | 84 | 12 | 87.5% |
+| ✓ `src\autosar_pdf2txt\models\enums.py` | 10 | 10 | 0 | 100.0% |
+| ✓ `src\autosar_pdf2txt\models\types.py` | 51 | 51 | 0 | 100.0% |
 | ✓ `src\autosar_pdf2txt\parser\__init__.py` | 2 | 2 | 0 | 100.0% |
-| ✗ `src\autosar_pdf2txt\parser\pdf_parser.py` | 413 | 357 | 56 | 86.4% |
+| ✗ `src\autosar_pdf2txt\parser\pdf_parser.py` | 413 | 401 | 12 | 97.1% |
 | ✓ `src\autosar_pdf2txt\writer\__init__.py` | 2 | 2 | 0 | 100.0% |
 | ✗ `src\autosar_pdf2txt\writer\markdown_writer.py` | 160 | 133 | 27 | 83.1% |
 
@@ -29,8 +33,8 @@ COVERAGE REPORT - MARKDOWN FORMAT
 | Source File | Coverage | Missing Statements |
 |-------------|----------|-------------------|
 | `src\autosar_pdf2txt\cli\autosar_cli.py` | 80.7% (67/83) | 16 stmts (19.3%) |
-| `src\autosar_pdf2txt\models\autosar_models.py` | 93.3% (168/180) | 12 stmts (6.7%) |
-| `src\autosar_pdf2txt\parser\pdf_parser.py` | 86.4% (357/413) | 56 stmts (13.6%) |
+| `src\autosar_pdf2txt\models\containers.py` | 87.5% (84/96) | 12 stmts (12.5%) |
+| `src\autosar_pdf2txt\parser\pdf_parser.py` | 97.1% (401/413) | 12 stmts (2.9%) |
 | `src\autosar_pdf2txt\writer\markdown_writer.py` | 83.1% (133/160) | 27 stmts (16.9%) |
 
 ======================================================================
